requirements.txt
aws-cdk-lib==2.31.0
constructs>=10.0.0,<11.0.0
PyYAML~=6.0
pip~=22.0.4
wheel~=0.37.1
setuptools~=60.5.0
boto3~=1.20.24
botocore~=1.23.24
